refactor(agent): isolate normalized resident author selection
